Extract PySamplerContext into its own module
