+++ rustuna_pyo3/src/sampler.rs
@@ -5,19 +5,20 @@ use pyo3::exceptions::PyRuntimeError;
 use pyo3::Py;
 use pyo3::{prelude::*, types::PyType};
 
-use rustuna_core::sampler::{Context as SamplerContext, Sampler};
+use rustuna_core::sampler::Sampler;
 use rustuna_core::storage::Storage;
 use rustuna_core::trial::TrialStateValues;
 use rustuna_sampler::tpe::{TpeConfig, TpeSampler};
 
 pub mod cmaes;
+mod context;
 pub mod python;
 pub mod random;
+pub use context::PySamplerContext;
 
 use crate::distribution::PyDistribution;
 use crate::pyobject_storage::PyPyObjectStorage;
 use crate::storage::PyStorage;
-use crate::study::PyDirection;
 use crate::trial::PyTrialState;
 
 #[derive(Clone)]
@@ -168,57 +169,4 @@ fn extract_storage(storage: Py<PyAny>) -> PyResult<Arc<RwLock<dyn Storage>>> {
             ))
         }
     })
-}
-
-#[derive(Clone)]
-#[pyclass(name = "SamplerContext")]
-#[pyo3(module = "rustuna")]
-pub struct PySamplerContext {
-    context: SamplerContext,
-}
-impl From<SamplerContext> for PySamplerContext {
-    fn from(item: SamplerContext) -> Self {
-        PySamplerContext { context: item }
-    }
-}
-#[allow(non_local_definitions)]
-#[pymethods]
-impl PySamplerContext {
-    #[new]
-    #[pyo3(signature = (*, study_id, trial_number, trial_id, directions))]
-    pub fn py_new(
-        study_id: u32,
-        trial_number: u32,
-        trial_id: u32,
-        directions: Vec<PyDirection>,
-    ) -> PyResult<Self> {
-        Ok(PySamplerContext {
-            context: SamplerContext {
-                study_id,
-                trial_number,
-                trial_id,
-                directions: directions.into_iter().map(|d| d.into()).collect(),
-            },
-        })
-    }
-    #[getter]
-    fn study_id(&self) -> u32 {
-        self.context.study_id
-    }
-    #[getter]
-    fn trial_number(&self) -> u32 {
-        self.context.trial_number
-    }
-    #[getter]
-    fn trial_id(&self) -> u32 {
-        self.context.trial_id
-    }
-    #[getter]
-    fn directions(&self) -> Vec<PyDirection> {
-        self.context
-            .directions
-            .iter()
-            .map(|d| d.clone().into())
-            .collect()
-    }
 }
